README.md
The 2016 Zika Virus outbreak affected several countries, most of them located in the Americas. It is extremely important to keep records on cases related to the outbreak in order to monitor the disease. Each country reported the cases using different methodologies so this project aims to get some insights on the development of the epidemic locally but also globally.